Rename authentication test files
Relates-to: OLPEDGE-594
