+++ olp-cpp-sdk-authentication/tests/AuthenticationTest.h
